# [SPARK-9926] [SQL] parallelize file listing for partitioned Hive table