--- a/core/src/main/scala/org/apache/spark/deploy/SparkHadoopUtil.scala
+++ b/core/src/main/scala/org/apache/spark/deploy/SparkHadoopUtil.scala
@@ -20,21 +20,19 @@ package org.apache.spark.deploy
 import java.io.{ByteArrayInputStream, DataInputStream}
 import java.lang.reflect.Method
 import java.security.PrivilegedExceptionAction
-import java.util.{Arrays, Comparator, Map}
+import java.util.{Arrays, Comparator}
 
 import scala.collection.JavaConversions._
 import scala.concurrent.duration._
 import scala.language.postfixOps
 import scala.util.control.NonFatal
 
-import com.google.common.collect.MapMaker
 import com.google.common.primitives.Longs
-
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.FileSystem.Statistics
 import org.apache.hadoop.fs.{FileStatus, FileSystem, Path, PathFilter}
 import org.apache.hadoop.hdfs.security.token.delegation.DelegationTokenIdentifier
-import org.apache.hadoop.mapred.{FileInputFormat, FileSplit, InputSplit, JobConf}
+import org.apache.hadoop.mapred.JobConf
 import org.apache.hadoop.mapreduce.JobContext
 import org.apache.hadoop.mapreduce.{TaskAttemptContext => MapReduceTaskAttemptContext}
 import org.apache.hadoop.mapreduce.{TaskAttemptID => MapReduceTaskAttemptID}
@@ -377,61 +375,6 @@ class SparkHadoopUtil extends Logging {
     val confKey = s"fs.${scheme}.impl.disable.cache"
     newConf.setBoolean(confKey, true)
     newConf
-  }
-
-  // This map holds input splits per Hive partition path.
-  private val inputSplitsCache: Map[String, Array[InputSplit]] = new MapMaker().makeMap()
-
-  /**
-   * Populate inputSplitsCache. Listing multiple partition paths is done in parallel, and
-   * input splits are grouped by their partition paths and stored in the map.
-   */
-  def computeInputSplits(
-      inputPaths: Seq[String],
-      jobConf: JobConf,
-      inputFormat: FileInputFormat[_, _],
-      minSplits: Int): Unit = {
-    addCredentials(jobConf)
-    val inputSplits = inputFormat.getSplits(jobConf, minSplits)
-    val groupedInputSplits = inputSplits.groupBy(_.asInstanceOf[FileSplit].getPath.getParent)
-    inputPaths.foreach { path =>
-      var files: Array[InputSplit] = Array()
-      groupedInputSplits.foreach { case (commonParentPath, array) =>
-        if (commonParentPath.toString.contains(path)) {
-          files = Array.concat(files, array)
-        }
-      }
-      if (files.nonEmpty) {
-        inputSplitsCache.put(path, files)
-      }
-    }
-  }
-
-  /**
-   * Return true if input splits for all the given partition paths exist, or false.
-   */
-  def hasInputSplitsCache(paths: Seq[String]): Boolean = {
-    inputSplitsCache.nonEmpty && paths.forall(inputSplitsCache.containsKey(_))
-  }
-
-  /**
-   * Return cached input splits for all the given partition paths, or nothing.
-   */
-  def getInputSplitsCache(paths: Seq[String]): Array[InputSplit] = {
-    var cache: Array[InputSplit] = Array()
-    if (hasInputSplitsCache(paths)) {
-      for (path <- paths) {
-        cache = Array.concat(cache, inputSplitsCache.get(path))
-      }
-    }
-    cache
-  }
-
-  /**
-   * Delete all the cached input splits
-   */
-  def clearInputSplitsCache(): Unit = {
-    inputSplitsCache.clear
   }
 }