--- a/core/src/main/scala/org/apache/spark/rdd/HadoopRDD.scala
+++ b/core/src/main/scala/org/apache/spark/rdd/HadoopRDD.scala
@@ -35,7 +35,6 @@ import org.apache.hadoop.mapred.Reporter
 import org.apache.hadoop.mapred.JobID
 import org.apache.hadoop.mapred.TaskAttemptID
 import org.apache.hadoop.mapred.TaskID
-import org.apache.hadoop.mapred.FileInputFormat
 import org.apache.hadoop.mapred.lib.CombineFileSplit
 import org.apache.hadoop.util.ReflectionUtils
 
@@ -106,7 +105,8 @@ class HadoopRDD[K, V](
     inputFormatClass: Class[_ <: InputFormat[K, V]],
     keyClass: Class[K],
     valueClass: Class[V],
-    minPartitions: Int)
+    minPartitions: Int,
+    @transient inputSplitsCache: Option[Array[InputSplit]] = None)
   extends RDD[(K, V)](sc, Nil) with Logging {
 
   if (initLocalJobConfFuncOpt.isDefined) {
@@ -198,21 +198,17 @@ class HadoopRDD[K, V](
   }
 
   override def getPartitions: Array[Partition] = {
-    val jobConf = getJobConf()
-    val inputPaths = FileInputFormat.getInputPaths(jobConf).map(_.toString)
-    val inputSplits: Array[InputSplit] =
-      // first check whether input splits are already computed
-      if (SparkHadoopUtil.get.hasInputSplitsCache(inputPaths)) {
-        SparkHadoopUtil.get.getInputSplitsCache(inputPaths)
-      } else {
-        // add the credentials here as this can be called before SparkContext initialized
-        SparkHadoopUtil.get.addCredentials(jobConf)
-        val inputFormat = getInputFormat(jobConf)
-        if (inputFormat.isInstanceOf[Configurable]) {
-          inputFormat.asInstanceOf[Configurable].setConf(jobConf)
-        }
-        inputFormat.getSplits(jobConf, minPartitions)
+    // First check whether input splits are already computed
+    val inputSplits: Array[InputSplit] = inputSplitsCache.getOrElse {
+      val jobConf = getJobConf()
+      // Add the credentials here as this can be called before SparkContext initialized
+      SparkHadoopUtil.get.addCredentials(jobConf)
+      val inputFormat = getInputFormat(jobConf)
+      if (inputFormat.isInstanceOf[Configurable]) {
+        inputFormat.asInstanceOf[Configurable].setConf(jobConf)
       }
+      inputFormat.getSplits(jobConf, minPartitions)
+    }
     val array = new Array[Partition](inputSplits.size)
     for (i <- 0 until inputSplits.size) {
       array(i) = new HadoopPartition(id, i, inputSplits(i))