--- a/sql/hive/src/main/scala/org/apache/spark/sql/hive/TableReader.scala
+++ b/sql/hive/src/main/scala/org/apache/spark/sql/hive/TableReader.scala
@@ -28,9 +28,11 @@ import org.apache.hadoop.hive.serde2.objectinspector.primitive._
 import org.apache.hadoop.hive.serde2.objectinspector.{ObjectInspectorConverters, StructObjectInspector}
 import org.apache.hadoop.io.Writable
 import org.apache.hadoop.mapred.{FileInputFormat, InputFormat, JobConf}
+import org.apache.hadoop.util.ReflectionUtils
 
 import org.apache.spark.Logging
 import org.apache.spark.broadcast.Broadcast
+import org.apache.spark.deploy.SparkHadoopUtil
 import org.apache.spark.rdd.{EmptyRDD, HadoopRDD, RDD, UnionRDD}
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions._
@@ -182,8 +184,9 @@ class HadoopTableReader(
       }
     }
 
-    val hivePartitionRDDs = verifyPartitionPath(partitionToDeserializer)
-      .map { case (partition, partDeserializer) =>
+    val hivePartitions = verifyPartitionPath(partitionToDeserializer)
+
+    val hivePartitionRDDs = hivePartitions.map { case (partition, partDeserializer) =>
       val partDesc = Utilities.getPartitionDesc(partition)
       val partPath = partition.getDataLocation
       val inputPathStr = applyFilterIfNeeded(partPath, filterOpt)
@@ -240,6 +243,28 @@ class HadoopTableReader(
       }
     }.toSeq
 
+    // Compute input splits for all the partitions together if they have the same input format.
+    // This is faster than computing them individually because listing multiple input dirs can be
+    // done in parallel using "mapreduce.input.fileinputformat.list-status.num-threads".
+    val homogeneousInputFormat =
+      if (hivePartitions.groupBy { case (part, _) => part.getInputFormatClass }.size == 1) {
+        true
+      } else {
+        false
+      }
+
+    if (homogeneousInputFormat) {
+      val combinedInputPath = hivePartitions.map { case (part, _) => part.getLocation }.toSeq
+      val jobConf = new JobConf(hiveExtraConf)
+      val minPartitions = _minSplitsPerRDD * hivePartitions.size
+      val inputFormatClass = hivePartitions.head._1.getInputFormatClass
+      HadoopTableReader.initializeLocalJobConfFunc(combinedInputPath, relation.tableDesc)(jobConf)
+      val inputFormat =
+        ReflectionUtils.newInstance(inputFormatClass.asInstanceOf[Class[_]], jobConf)
+          .asInstanceOf[FileInputFormat[_, _]]
+      SparkHadoopUtil.get.computeInputSplits(combinedInputPath, jobConf, inputFormat, minPartitions)
+    }
+
     // Even if we don't use any partitions, we still need an empty RDD
     if (hivePartitionRDDs.size == 0) {
       new EmptyRDD[InternalRow](sc.sparkContext)
@@ -271,7 +296,7 @@ class HadoopTableReader(
     path: String,
     inputFormatClass: Class[InputFormat[Writable, Writable]]): RDD[Writable] = {
 
-    val initializeJobConfFunc = HadoopTableReader.initializeLocalJobConfFunc(path, tableDesc) _
+    val initializeJobConfFunc = HadoopTableReader.initializeLocalJobConfFunc(Seq(path), tableDesc) _
 
     val rdd = new HadoopRDD(
       sc.sparkContext,
@@ -292,8 +317,8 @@ private[hive] object HadoopTableReader extends HiveInspectors with Logging {
    * Curried. After given an argument for 'path', the resulting JobConf => Unit closure is used to
    * instantiate a HadoopRDD.
    */
-  def initializeLocalJobConfFunc(path: String, tableDesc: TableDesc)(jobConf: JobConf) {
-    FileInputFormat.setInputPaths(jobConf, Seq[Path](new Path(path)): _*)
+  def initializeLocalJobConfFunc(path: Seq[String], tableDesc: TableDesc)(jobConf: JobConf) {
+    FileInputFormat.setInputPaths(jobConf, path.map(new Path(_)): _*)
     if (tableDesc != null) {
       PlanUtils.configureInputJobPropertiesForStorageHandler(tableDesc)
       Utilities.copyTableJobPropertiesToConf(tableDesc, jobConf)