Instead of listing files on a per-partition basis, listing all the partitions together in parallel improves query performance quite a bit.

Here are benchmark results of my test query in the [jira](https://issues.apache.org/jira/browse/SPARK-9926) w/ setting ``mapreduce.input.fileinputformat.list-status.num-threads`` to 25-
* Parquet-backed Hive table
* Partitioned by dateint and hour
* Stored on S3

\# of files|\# of partitions|before|after|improvement|
------------|-------------------|---------|------|-----
972 | 1 | 38 secs | 20 secs | 1.9x
13646 | 24 | 354  secs | 28 secs | 12x
136222 | 240 | 3507  secs | 156 secs | 22x

The changes include-
* Added a new property ``spark.sql.hive.parallelFileListing`` that enables parallel file listing in a partitioned Hive table backed by Hadoop storage such as HDFS and S3. By default, false.
* When the property is true, ``TableReader`` computes input splits for all the partitions in parallel and cache them in ``HadoopRDD``.
* In ``HadoopRDD.getPartitions()``, first check whether input splits are already computed, and if so, reuse them.