@piaozhexiu 2015-08-18
I think unit tests time out because multiple test cases concurrently access `inputSplitsCache` that I introduced in `SparkHadoopUtil`. I will convert that central cache into local cache in `HadoopRDDs`. Hopefully that will fix the time out.