@piaozhexiu 2015-08-30
This PR is superseded by #8512. Closing.